+++ packages/react-core/src/filters/FilterQueryBuilder.js
@@ -1,6 +1,7 @@
 export const FilterTypes = Object.freeze({
   IN: 'in',
-  BETWEEN: 'between',
+  BETWEEN: 'between', // [a, b] both are included
+  CLOSED_OPEN: 'closed_open', // [a, b) a is included, b is not
   TIME: 'time'
 });