+++ packages/react-ui/src/widgets/WrapperWidgetUI.js
@@ -61,6 +61,7 @@ const useStyles = makeStyles((theme) => ({
     flex: 1,
     padding: 0,
     alignItems: 'flex-start',
+    justifyContent: 'flex-start',
     cursor: (props) => (props.expandable ? 'pointer' : 'default'),
     '& .MuiButton-label': {
       ...theme.typography.body1,