+++ packages/react-widgets/src/widgets/PieWidget.js
@@ -19,7 +19,9 @@ import { selectIsViewportFeaturesReadyForSource } from '@carto/react-redux/';
  * @param  {string} props.operation - Operation to apply to the operationColumn. Must be one of those defined in `AggregationTypes` object.
  * @param  {Function} [props.formatter] - Function to format the value that appears in the tooltip.
  * @param  {Function} [props.tooltipFormatter] - Function to return the HTML of the tooltip.
+ * @param  {object} props.labels - Object that maps category name with a chosen label
  * @param  {string} props.height - Height of the chart
+ * @param  {boolean} [props.animation] - Enable/disable widget animations on data updates. Enabled by default.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  * @param  {Object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]()
@@ -34,6 +36,8 @@ function PieWidget({
   operation,
   formatter,
   tooltipFormatter,
+  labels,
+  animation,
   colors,
   onError,
   wrapperProps,
@@ -119,6 +123,8 @@ function PieWidget({
           height={height}
           tooltipFormatter={tooltipFormatter}
           colors={colors}
+          labels={labels}
+          animation={animation}
           selectedCategories={selectedCategories}
           onSelectedCategoriesChange={handleSelectedCategoriesChange}
         />
@@ -139,6 +145,8 @@ PieWidget.propTypes = {
   operation: PropTypes.oneOf(Object.values(AggregationTypes)).isRequired,
   formatter: PropTypes.func,
   tooltipFormatter: PropTypes.func,
+  labels: PropTypes.object,
+  animation: PropTypes.bool,
   onError: PropTypes.func,
   colors: PropTypes.arrayOf(PropTypes.string),
   wrapperProps: PropTypes.object,
@@ -146,6 +154,7 @@ PieWidget.propTypes = {
 };
 
 PieWidget.defaultProps = {
+  animation: true,
   wrapperProps: {},
   noDataAlertProps: {}
 };